fix: add client

+++ src/core/controller/Controller.js
@@ -18,17 +18,19 @@ const Puppy = require('../Puppy')
 // }
 
 class Controller {
-  constructor(entity, params = {}) {
-    this.model = Puppy.models[entity]
-    this.templates = Puppy.templates[entity]
-
-    this.find = find.bind(this)
-    this.findOne = findOne.bind(this)
-    this.create = create.bind(this)
-    this.update = update.bind(this)
-    this.deleteOne = deleteOne.bind(this)
-    this.print = print.bind(this)
-    this.email = email.bind(this)
+  constructor(params = {}, entity = null) {
+    if (entity) {
+      this.model = Puppy.models[entity]
+      this.templates = Puppy.templates[entity]
+  
+      this.find = find.bind(this)
+      this.findOne = findOne.bind(this)
+      this.create = create.bind(this)
+      this.update = update.bind(this)
+      this.deleteOne = deleteOne.bind(this)
+      this.print = print.bind(this)
+      this.email = email.bind(this)
+    }
 
     Object.keys(params).forEach(key => {
       this[key] = params[key].bind(this)
